Update licensing.md - fix anchors
Link anchors with capital letters (like in https://opensource.zalando.com/docs/resources/licensing/#Code-licensed-under-AGPL-or-variants-thereof) are accessible in GitHub, but do not work properly on the website.
While this could be solved on the code level by fixing URL dispatching logic, I am not sure I'd be able to tackle this.
Simple solution for this page - to fix anchors on this page. Lowercased links both working on GitHub and on a static-generated website.

Signed-off-by: Rust Saiargaliev <rustem.saiargaliev@thermondo.de>